Refactor code, split in smaller components.
Navbar is now splitted in smaller components, since the bootstrap
made it look obfuscated with huge amount of line codes.

For the Landing page, there's new component <Card /> - which takes
data in the props. It contains src for image, title and text.

RegisterPage was tested for errors, soon changes will be made to
login page also.