Recommender server is up
- Created 2 content-based systems, they are differ if it comes about
weighting words in documents.

1. CV_CB (CountVectorizer_ContentBased) - count words and don't set them
any weight.

2. TF_IDF_CB (TermFrequency_InverseDocumentFrequency_ContentBased) - the
more word appear in one description and less in others - the more weight
it gets. If there's a word that doesn't matter (for example it is used
in many other documents) it doesn't have high weigh.
For example, word: "frequency" will be used more likely in scientific
books, magazines - so since common books won't have such a word, for
algorithm it means, that books which have "frequency" word in them
(especially mutliple times), should be more similar than those, which
don't have that word.